+++ cinelerra/savefile.C
@@ -110,7 +110,10 @@ int Save::handle_event()
 		else
 		{
 			char string[BCTEXTLEN];
-			sprintf(string, _("\"%s\" %dC written"), mwindow->session->filename, strlen(file.string));
+			sprintf(string, 
+				_("\"%s\" %dC written"), 
+				mwindow->session->filename, 
+				(int)strlen(file.string));
 			mwindow->gui->show_message(string);
 		}
 		mwindow->session->changes_made = 0;
@@ -216,7 +219,7 @@ void SaveAs::run()
 	else
 	{
 		char string[BCTEXTLEN];
-		sprintf(string, _("\"%s\" %dC written"), filename, strlen(file.string));
+		sprintf(string, _("\"%s\" %dC written"), filename, (int)strlen(file.string));
 		mwindow->gui->lock_window("SaveAs::run 2");
 		mwindow->gui->show_message(string);
 		mwindow->gui->unlock_window();
